refactor(frontend): update nursing management components and backend entities
- Remove description field from level query parameters and table display
- Add remark and updateTime columns to level and plan tables
- Remove sortNo and status fields from plan query parameters
- Simplify project component structure with cleaner template layout
- Add status dropdown filter and proper status display in project table
- Update NursingLevel entity to use Long ID and update date
- Change NursingLevel lplanId from Integer to Long
- Update NursingPlan entity to use Long ID and sortNo with updated date
- Remove unused service interfaces and controller classes
- Adjust level menu parent ID in database script
- Remove description condition from level mapper query
- Update mapper XML parameter types from Integer to Long
- Remove unused service implementation classes

+++ YiSmartCloud-background/YiSmartCloud-nursing-platform/src/main/java/org/FlyingSparrow/YiSmartCloud/serve/domain/NursingLevel.java
@@ -9,14 +9,14 @@ import lombok.Data;
  * 护理等级对象 nursing_level
  * 
  * @author Siborne
- * @date 2026-03-02
+ * @date 2026-03-23
  */
 @Data
 public class NursingLevel extends BaseEntity {
     private static final long serialVersionUID = 1L;
 
     /** 主键ID */
-    private Integer id;
+    private Long id;
 
     /** 等级名称 */
     @Excel(name = "等级名称")
@@ -24,7 +24,7 @@ public class NursingLevel extends BaseEntity {
 
     /** 护理计划ID */
     @Excel(name = "护理计划ID")
-    private Integer lplanId;
+    private Long lplanId;
 
     /** 护理费用 */
     @Excel(name = "护理费用")
